--- ui/src/scenes/Home/components/NeedsContainer/components/NeedsList/components/MissingRealizersAmount/index.js
@@ -1,3 +1,0 @@
-import MissingRealizersAmount from './MissingRealizersAmount';
-
-export default MissingRealizersAmount;
