+++ ui/src/scenes/Home/components/NeedsContainer/components/NeedsList/components/NeedsListItem/components/MissingRealitiesOnNeed/index.js
@@ -1,0 +1,3 @@
+import MissingRealitiesOnNeed from './MissingRealitiesOnNeed';
+
+export default MissingRealitiesOnNeed;
